+++ tests/testthat/test_clone.R
@@ -144,3 +144,192 @@ test_that("Test assigning clones works for heavy-only sc data", {
                                locus = "locus", nproc=1),
                    "Single cell mode requested, but")
 })
+
+
+#### Single cell 
+
+test_that("Test hierarchicalClones light chain split works", {
+    
+    db <- data.frame(
+        sequence_id=c("seq1","seq2","seq3","seq4","seq5","seq6","seq7","seq8"),
+        cell_id=c("cell1","cell1","cell2","cell2","cell3","cell3","cell4","cell4"),
+        v_call=c("IGHV1*01","IGLV1*01","IGHV1*01","IGLV1*01","IGHV1*01","IGLV2*01", "IGHV3*01,IGHV1*01","IGLV1*01"),
+        d_call=c("IGHD1*01",NA,"IGHD1*01",NA,"IGHD1*01",NA,"IGHD1*01",NA),
+        j_call=c("IGHJ1*01","IGLJ1*01","IGHJ1*01","IGLJ1*01*01","IGHJ1*01","IGLJ1*01","IGHJ1*01","IGLJ1*01"),
+        junction=c("TCGAAATTC","TCGTTTTTC","TCGAAATTC","TCGTTTTTTTTC","TCGAAATTC","TCGTTTTTC","TCGAAATTC","TCGTTTTTC")
+    )
+    db$chain <- "light"
+    db$chain[grepl("IGH",db[['v_call']])] <- "heavy"
+    db$locus <- alakazam::getLocus(db$v_call)
+    db$junction_len <- stringi::stri_length(db[['junction']])  
+
+    # prepare_db uses groupGenes to find vj gene groups, but 
+    # uses junc_len = NULL
+    # It adds the junction length groups outside groupGenes
+    # requires both cell_id and locus
+    # scoper:::prepare_db
+    db_only_heavy_T_locus <- scoper:::prepare_db(db,only_heavy = T, 
+                                                      cell_id="cell_id",
+                                                      locus="locus", first=F)$db
+    expect_true(all(db_only_heavy_T_locus$vj_group =="G1"))
+    
+    # scoper:::prepare_db
+    db_only_heavy_F_locus_first_T <- scoper:::prepare_db(db,only_heavy = F, 
+                                           cell_id="cell_id",
+                                           locus="locus",
+                                           first=T)$db
+    expect_equal(db_only_heavy_F_locus_first_T$vj_group, c("G1","G1","G1","G1","G2","G2","G3","G3"))
+    
+    db_only_heavy_F_locus_first_F <- scoper:::prepare_db(db,only_heavy = F, 
+                                                       cell_id="cell_id",
+                                                       locus="locus",
+                                                       first=F)$db
+    expect_equal(db_only_heavy_F_locus_first_F$vj_group, c("G1","G1","G1","G1","G2","G2","G1","G1"))
+    
+    
+    ## Test hierachicalClones
+    ## case only_heavy   split_light first
+    ## 1    T            F           F
+    ## 2    T            F           T
+    ## 3    T            T           F
+    ## 4    T            T           T
+    ## 5    F            F           F
+    ## 6    F            F           T
+    ## 7    F            T           F
+    ## 8    F            T           T
+    
+    # Case 1
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=FALSE,
+        first=F, # default is first=F
+        nproc=1)
+    expect_true(all(clones@db[['clone_id']] == "1"))
+    
+    # Case 2
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=FALSE,
+        first=T, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']],c("1","1","1","1","1","1","2","2"))
+    
+    # Case 3
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=TRUE,
+        first=F, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']], c("1","1","2","2","2","2","3","3"))
+    
+    # Case 4
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=TRUE,
+        first=T, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']], c("1","1","2","2","3","3","4","4"))
+    
+    # Case 5
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=FALSE,
+        split_light=FALSE,
+        first=F, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']],c("1","1","1","1","1","1","2","2"))    
+    
+    # Case 6
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=FALSE,
+        split_light=FALSE,
+        first=T, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']],c("1","1","1","1","2","2","3","3"))    
+    
+    # Case 7
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=FALSE,
+        split_light=TRUE,
+        first=F, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']],c("1","1","2","2","2","2","3","3"))    
+    
+    # Case 8
+    clones <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=FALSE,
+        split_light=TRUE,
+        first=T, # default is first=F
+        nproc=1)
+    expect_equal(clones@db[['clone_id']],c("1","1","2","2","3","3","4","4"))      
+    
+    ## Wwhat happens with the second groupGenes (inside the light chain split) when
+    # first=false, but the "linker" ambiguous call was left out of the same cluster id
+    # because of the distance threshold? This groupGenes could be splitting again by vj calls...
+    # seq2 is the linker, and the juction is one nt different. Everything else is the same.
+    db <- data.frame(
+        sequence_id=c("seq1","seq2","seq3","seq4","seq5","seq6"),
+        cell_id=c("1","2","3","1","2","3"),
+        v_call=c("IGHV1", "IGHV1,IGHV2","IGHV2","IGLV1","IGLV1","IGLV1"),
+        j_call=c("IGHJ1","IGHJ1","IGHJ1","IGHJ1","IGHJ1","IGHJ1"),
+        junction=c("TCGAAATTC","TCGAACTTC","TCGAAATTC","TCGAAATTC","TCGAAATTC","TCGAAATTC")
+    )
+    db$locus <- alakazam::getLocus(db$v_call)
+    db
+    clones_split_F <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=FALSE,
+        first=F, # default is first=F
+        nproc=1)
+    
+    clones_split_T <- hierarchicalClones(
+        db,
+        threshold=0,
+        cell_id='cell_id',
+        locus='locus',
+        only_heavy=TRUE,
+        split_light=TRUE,
+        first=F, # default is first=F
+        nproc=1)
+    # expecting same results because the light chains are the same.
+    expect_equal(clones_split_F@db[['clone_id']], clones_split_T@db[['clone_id']])
+    
+    ## TODO: multiple light chains?
+    
+    ## TODO: summaries etc are using clone ids created before the light chain split?
+})